models-library: grouped settings
