+++ packages/models-library/src/models_library/basic_types.py
@@ -1,0 +1,16 @@
+
+from enum import Enum
+
+
+from pydantic import conint, constr
+
+PortInt = conint(gt=0, lt=65535)
+VersionTag = constr(regex=r"^v\d$")
+
+
+
+class LogLevel(str, Enum):
+    DEBUG = "DEBUG"
+    INFO = "INFO"
+    WARNING = "WARNING"
+    ERROR = "ERROR"
