+++ packages/models-library/src/models_library/settings/__init__.py
@@ -1,0 +1,3 @@
+"""
+    Common settings
+"""
