+++ packages/models-library/src/models_library/settings/celery.py
