+++ packages/models-library/src/models_library/settings/postgres.py
@@ -1,21 +1,11 @@
-from enum import Enum
-
 # pylint: disable=no-self-argument
 # pylint: disable=no-self-use
+
+from enum import Enum
 from typing import Optional
 
 from pydantic import BaseSettings, PostgresDsn, SecretStr, conint, constr, validator
-
-PortInt = conint(gt=0, lt=65535)
-VersionTag = constr(regex=r"^v\d$")
-
-
-class LogLevel(str, Enum):
-    DEBUG = "DEBUG"
-    INFO = "INFO"
-    WARNING = "WARNING"
-    ERROR = "ERROR"
-
+from ..basic_types import PortInt, VersionTag, LogLevel
 
 class PostgresSettings(BaseSettings):
     dns: Optional[PostgresDsn] = None
