+++ packages/models-library/src/models_library/settings/rabbit.py
